[FIX] 오류 수정

+++ MiWorkDas/app/src/main/java/com/midas/miworkdas/model/response/WorkEnd.kt
@@ -3,5 +3,5 @@ package com.midas.miworkdas.model.response
 data class WorkEnd(
     val success: Boolean,
     val memberWeekHour: Int,
-    val memberWorkTime: Double,
+    val memberWorkTime: Int,
 )
